cleanup ints

+++ Controller Classes/EndpointTableDataSource.h
@@ -13,17 +13,17 @@
 
 - (void)setEndpointArray:(NSMutableArray*)newEndpointArray;
 
-- (int)numberOfRowsInTableView:(NSTableView*)tableView;
-- (id)tableView:(NSTableView*)tableView objectValueForTableColumn:(NSTableColumn*)column row:(int)rowIndex;
+- (NSInteger)numberOfRowsInTableView:(NSTableView*)tableView;
+- (id)tableView:(NSTableView*)tableView objectValueForTableColumn:(NSTableColumn*)column row:(NSInteger)rowIndex;
 - (BOOL)control:(NSControl*)control isValidObject:(id)value;
-- (void)tableView:(NSTableView*)tableView setObjectValue:(id)value forTableColumn:(NSTableColumn*)column row:(int)rowIndex;
+- (void)tableView:(NSTableView*)tableView setObjectValue:(id)value forTableColumn:(NSTableColumn*)column row:(NSInteger)rowIndex;
 - (void)deleteSelection:(NSTableView*)tableView;
 
 - (BOOL)tabView:(NSTabView*)tabView shouldSelectTabViewItem:(NSTabViewItem*)tabViewItem;
 
 - (void)tableView:(NSTableView*)tableView newEndpointWithName:(NSString*)name;
-- (void)tableView:(NSTableView*)tableView addEndpoint:(PYMIDIVirtualEndpoint*)endpoint atIndex:(int)index;
-- (void)tableView:(NSTableView*)tableView removeEndpointAtIndex:(int)index;
-- (void)tableView:(NSTableView*)tableView setName:(NSString*)name forEndpointAtIndex:(int)index;
+- (void)tableView:(NSTableView*)tableView addEndpoint:(PYMIDIVirtualEndpoint*)endpoint atIndex:(NSInteger)index;
+- (void)tableView:(NSTableView*)tableView removeEndpointAtIndex:(NSInteger)index;
+- (void)tableView:(NSTableView*)tableView setName:(NSString*)name forEndpointAtIndex:(NSInteger)index;
 
 @end
