+++ PYMIDI/PYMIDIManager.m
@@ -143,7 +143,7 @@ midiNotifyProc (const MIDINotification* message, void* refCon)
     
     // Find any non-virtual endpoints that we don't already know about
     int i;
-    int count = MIDIGetNumberOfSources();
+    NSUInteger count = MIDIGetNumberOfSources();
     for (i = 0; i < count; i++) {
         MIDIEndpointRef midiEndpointRef = MIDIGetSource (i);
         
@@ -220,7 +220,7 @@ midiNotifyProc (const MIDINotification* message, void* refCon)
     
     // Find any non-virtual endpoints that we don't already know about
     int i;
-    int count = MIDIGetNumberOfDestinations();
+    NSUInteger count = MIDIGetNumberOfDestinations();
     for (i = 0; i < count; i++) {
         MIDIEndpointRef midiEndpointRef = MIDIGetDestination (i);
